set up dummy test build
