+++ platformio.ini
@@ -61,3 +61,10 @@ board_build.variant = quail
 build_flags =
  -DQUAIL
 
+[env:dummy]
+src_filter = +<dummy/>
+board = quail
+board_build.variant = quail
+build_flags =
+ -DDUMMY
+
