+++ src/dummy/System.hpp
@@ -1,0 +1,20 @@
+#pragma once
+
+class System;
+
+#include <Arduino.h>
+
+#include "SPI.h"
+
+class System
+{ 
+public:
+    
+    class Tasks
+    {
+    public:
+    };
+    Tasks tasks;
+};
+
+#include "main.hpp"
